feat: BankAccountController 단위 테스트

+++ hec-core/src/main/java/com/example/heccore/bank/model/BankAccountWithUserNameVO.java
@@ -15,7 +15,7 @@ public class BankAccountWithUserNameVO extends BaseVO {
     private Long userId;
     private Bank bank;
     private Long accountNumber;
-    private BigDecimal balance;
+    private int balance;
     private boolean isDeleted;
     private String name; // 사용자 이름 추가
