+++ hec-mybatis/src/main/java/com/example/hecmybatis/bankAccount/dto/request/BankAccountAmountRequestDto.java
@@ -2,6 +2,6 @@ package com.example.hecmybatis.bankAccount.dto.request;
 
 import java.math.BigDecimal;
 
-public record BankAccountAmountRequestDto(BigDecimal amount) {
+public record BankAccountAmountRequestDto(int amount) {
 
 }
