+++ hec-mybatis/src/main/java/com/example/hecmybatis/bankAccount/dto/request/BankAccountRequestDto.java
@@ -3,6 +3,6 @@ package com.example.hecmybatis.bankAccount.dto.request;
 import com.example.heccore.common.enums.Bank;
 import java.math.BigDecimal;
 
-public record BankAccountRequestDto(Long userId, Bank bank, BigDecimal balance) {
+public record BankAccountRequestDto(Long userId, Bank bank, int balance) {
 
 }
